Kompletna zmiana z folderami, inne polozene. Wykonane migracje
