+++ sklep/admin.py
@@ -5,6 +5,7 @@ from .models import Category, Product
 class CategoryAdmin(admin.ModelAdmin):
  list_display = ['name', 'slug']
  prepopulated_fields = {'slug': ('name',)}
+
 @admin.register(Product)
 class ProductAdmin(admin.ModelAdmin):
  list_display = ['name', 'slug', 'price',
