+++ sklep/templates/sklep/produkty/szczegoly.html
@@ -3,6 +3,7 @@
 
 {% block title %}
   {{ product.name }}
+<!-- {% if category %}{{ category.title }}{% else %}Produkty{% endif %} -->
 {% endblock %}
 
 {% block content %}
