+++ sklep/urls.py
@@ -1,0 +1,10 @@
+from django.urls import path
+from . import views
+
+app_name = 'sklep'
+
+urlpatterns = [
+    path('', views.product_list, name='product_list'),
+    path('<slug:category_slug>/', views.product_list, name='product_list_by_category'),
+    path('<int:id>/<slug:slug>/', views.product_detail, name='product_detail'),
+]
